Splited ProcessMonitor and ProcessInfo
