+++ src/UI/main.cpp
@@ -3,6 +3,7 @@
 #include "processesmodel.h"
 #include "processmonitor.h"
 #include "processesseeker.h"
+
 int main(int argc, char *argv[])
 {
     QApplication a(argc, argv);
