+++ src/UI/processmonitor.h
@@ -2,36 +2,9 @@
 #include <QString>
 #include <vector>
 #include <windows.h>
+#include "processinfo.h"
 
 class IProcessesSeeker;
-
-struct ProcessInfo{
-    ProcessInfo(DWORD pid,const QString& Path);
-    DWORD Pid;
-    QString Path;
-    QString Name;
-    bool readPermission = true;
-    bool writePermission = true;
-    bool openPermission = true;
-    bool deletePermission = true;
-    bool isMonitored = false;
-    bool isDllInjected = false;
-
-    bool settingsEquals(const ProcessInfo& other) const;
-    friend bool operator<(const ProcessInfo& first , const ProcessInfo& second);
-    friend bool operator==(const ProcessInfo& first , const ProcessInfo& second);
-private:
-    QString getNameFromThePath();
-};
-
-enum class ProcessEditableFields{
-    readPerm,
-    writePerm,
-    openPerm,
-    deletePerm,
-    isMonitored,
-    isDllInjected
-};
 
 class IProcessMonitor{
 public:
